+++ lib/Payever/Payments/Http/RequestEntity/ClaimPaymentRequest.php
@@ -18,152 +18,64 @@ use Payever\Sdk\Core\Http\RequestEntity;
 /**
  * This class represents Claim Payment RequestInterface Entity
  *
- * @method ClaimContextInfoEntity     getClaimContextInfo()
- * @method ClaimDocumentsInfoEntity[] getClaimDocumentsInfo()
- *
- * @SuppressWarnings(PHPMD.LongVariable)
+ * @method bool   getIsNonInclusive()
+ * @method bool   getIsLegal()
+ * @method bool   getIsDisputed()
+ * @method bool   getIsGuaranteeExisting()
+ * @method int    getPolicyId()
+ * @method string getBusinessUnitCode()
+ * @method string getExtensionId()
+ * @method self   setIsNonInclusive(bool $isNonInclusive)
+ * @method self   setIsLegal(bool $isLegal)
+ * @method self   setIsDisputed(bool $isDisputed)
+ * @method self   setIsGuaranteeExisting(bool $isGuaranteeExisting)
+ * @method self   setPolicyId(string $policyId)
+ * @method self   setBusinessUnitCode(string $businessUnitCode)
+ * @method self   setExtensionId(string $extensionId)
  */
 class ClaimPaymentRequest extends RequestEntity
 {
-    const UNDERSCORE_ON_SERIALIZATION = false;
+    /** @var bool $isNonInclusive */
+    protected $isNonInclusive = false;
 
-    /** @var ClaimContextInfoEntity $claimContextInfo */
-    protected $claimContextInfo;
+    /** @var bool $isLegal */
+    protected $isLegal = false;
 
-    /** @var ClaimDocumentsInfoEntity[] $claimDocumentsInfo */
-    protected $claimDocumentsInfo = [];
+    /** @var bool $isDisputed */
+    protected $isDisputed = false;
 
-    /**
-     * {@inheritdoc}
-     */
-    public function __construct($data = null)
-    {
-        $this->claimContextInfo = new ClaimContextInfoEntity();
+    /** @var bool $isGuaranteeExisting */
+    protected $isGuaranteeExisting = false;
 
-        parent::__construct($data);
-    }
+    /** @var string $policyId */
+    protected $policyId;
 
-    /**
-     * Sets claim context info
-     *
-     * @param ClaimContextInfoEntity|string $claimContextInfo
-     *
-     * @return $this
-     */
-    public function setClaimContextInfo($claimContextInfo)
-    {
-        if (!$claimContextInfo) {
-            return $this;
-        }
+    /** @var string $businessUnitCode */
+    protected $businessUnitCode;
 
-        if (is_string($claimContextInfo)) {
-            $claimContextInfo = json_decode($claimContextInfo);
-        }
-
-        if (!is_array($claimContextInfo) && !is_object($claimContextInfo)) {
-            return $this;
-        }
-
-        $this->claimContextInfo = new ClaimContextInfoEntity($claimContextInfo);
-
-        return $this;
-    }
-
-    /**
-     * Sets claim documents info
-     *
-     * @param array|string $claimDocumentsInfo
-     *
-     * @return $this
-     */
-    public function setClaimDocumentsInfo($claimDocumentsInfo)
-    {
-        if (!$claimDocumentsInfo) {
-            return $this;
-        }
-
-        if (is_string($claimDocumentsInfo)) {
-            $claimDocumentsInfo = json_decode($claimDocumentsInfo);
-        }
-
-        if (!is_array($claimDocumentsInfo)) {
-            return $this;
-        }
-
-        $this->claimDocumentsInfo = [];
-
-        foreach ($claimDocumentsInfo as $item) {
-            $this->claimDocumentsInfo[] = new ClaimDocumentsInfoEntity($item);
-        }
-
-        return $this;
-    }
-
-    /**
-     * Add claim document info
-     *
-     * @param ClaimDocumentsInfoEntity $document
-     *
-     * @return $this
-     */
-    public function addClaimDocumentEntity($document)
-    {
-        if (!($document instanceof ClaimDocumentsInfoEntity)) {
-            return $this;
-        }
-
-        $this->claimDocumentsInfo[] = $document;
-
-        return $this;
-    }
-
-    /**
-     * Set context info insolvency proceeding
-     *
-     * @param bool $isInsolvencyProceeding
-     *
-     * @return $this
-     */
-    public function setIsInsolvencyProceeding($isInsolvencyProceeding)
-    {
-        $this->claimContextInfo->setIsInsolvencyProceeding($isInsolvencyProceeding);
-
-        return $this;
-    }
-
-    /**
-     * Set context info invoice disputed
-     *
-     * @param bool $isInvoiceDisputed
-     *
-     * @return $this
-     */
-    public function setIsInvoiceDisputed($isInvoiceDisputed)
-    {
-        $this->claimContextInfo->setIsInvoiceDisputed($isInvoiceDisputed);
-
-        return $this;
-    }
+    /** @var string $extensionId */
+    protected $extensionId;
 
     /**
      * {@inheritdoc}
      */
     public function isValid()
     {
-        if (!$this->claimDocumentsInfo || !is_array($this->claimDocumentsInfo)) {
-            return false;
-        }
+        return parent::isValid()
+            && is_bool($this->isNonInclusive)
+            && is_bool($this->isLegal)
+            && is_bool($this->isDisputed)
+            && is_bool($this->isGuaranteeExisting)
+            && (!$this->policyId || is_string($this->policyId))
+            && (!$this->businessUnitCode || is_string($this->businessUnitCode))
+            && (!$this->extensionId || is_string($this->extensionId));
+    }
 
-        if (!$this->claimContextInfo->isValid()) {
-            return false;
-        }
-
-        foreach ($this->claimDocumentsInfo as $item) {
-            if (!($item instanceof ClaimDocumentsInfoEntity) || !$item->isValid()) {
-                return false;
-            }
-        }
-
-        return parent::isValid();
+    /**
+     * {@inheritdoc}
+     */
+    public function toArray($object = null)
+    {
+        return $object ? get_object_vars($object) : get_object_vars($this);
     }
 }